Update setup.sh
Changed lines 71&72.
Corrected ortography, corrected"Receiver Altitude" and precised in which unit the altitude ist to ben entered.
I have assumed feet. 
Please correct if the altitude is to be entered in meters. 
That was badly ambiguous and might impact the precision of MLAT.
May I suggest to clarify this on your website?